(PUP-5428) Add a query string with the environment when using post method
This patch adds a query string with the current environment.
Indeed, when the length of the uri is greater than 1024, the GET method
is converted to a POST method. All parameters are then stored in the
BODY of the request.
Putting the environment in the URI allows loadbalancers to redirect requests
regarding the environment.
This is specific to Puppet 4.x. Using Puppet 3.x, the 1st field of the uri
was the environment.